requirements.txt
Django==4.2.5
psycopg2-binary==2.9.7
djangorestframework==3.14.0
django-filter==23.2
djoser==2.2.0
django-debug-toolbar==4.2.0
pytest==7.4.2
model-bakery==1.15.0
django-extensions==3.2.3
